Refactor: the Absence Type has been moved to domain layer as a business logic.
Description: Instead of using a list of strings, we have defined an enum with a helper class to return the list. This change brings two benefits. Firstly, the Absence type now belongs to the domain layer, ensuring better separation of concerns. Secondly, if new types are introduced, we can easily update the business logic in absence_type.dart, which propagates throughout the system.

Took 27 minutes